refactor(acp): inline broadcast metrics and unify chunk metric seam
Fold the single-consumer publication-metrics module into the publication
owner (a throttled debug line is the only output), and route both the
single and batch apply paths through one recordAppliedTextChunk helper so
their per-chunk metrics cannot drift apart.
